requirements.txt
python-telegram-bot[rate-limiter,job-queue]>=20.0
python-dotenv>=0.19.0
dateparser>=1.1.0